+++ src/Makefile
@@ -2,12 +2,14 @@
 # main vgmstream code
 #
 
-# automatically get all possible .o by finding all .c
-OBJECTS =
-
 #SRCS = $(wildcard **/*.c) #GNUMake 3.81?
 SRCS = $(wildcard *.c) $(wildcard */*.c) $(wildcard */*/*.c)
+
+# get all possible .o by finding all .c
 OBJECTS = $(patsubst %.c,%.o,$(SRCS))
+
+# in case of renamed files
+OBJECTS_CLEAN = $(wildcard *.o) $(wildcard */*.o) $(wildcard */*/*.o)
 
 
 libvgmstream.a: $(OBJECTS)
@@ -20,6 +22,6 @@ libvgmstream.so: $(OBJECTS)
 #	$(CC) $(CFLAGS) -M -o vgmstream-deps
 
 clean:
-	$(RMF) $(OBJECTS) libvgmstream.a
+	$(RMF) $(OBJECTS_CLEAN) libvgmstream.a
 
 .PHONY: clean
