+++ winamp/in_vgmstream.vcxproj.filters
@@ -21,6 +21,18 @@
     <ClInclude Include="resource.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="sdk\in2.h">
+      <Filter>sdk\Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="sdk\ipc_pe.h">
+      <Filter>sdk\Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="sdk\out.h">
+      <Filter>sdk\Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="sdk\wa_ipc.h">
+      <Filter>sdk\Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="resource.rc">
@@ -32,6 +44,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="in_streamfile.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="in_utils.c">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="in_vgmstream.c">
