+++ xmplay/xmp-vgmstream.vcxproj.filters
@@ -1,11 +1,26 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClCompile Include="xmp_vgmstream.c" />
+    <ClCompile Include="xmp_streamfile.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="xmp_utils.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="xmp_vgmstream.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="xmpfunc.h" />
-    <ClInclude Include="xmpin.h" />
+    <ClInclude Include="xmpfunc.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="xmpin.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="xmp_vgmstream.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="xmpin.def" />
